Adding new version of STAAR
Ok, so this is the new version of STAAR that I have written.  It is
approximately 3x-4x faster than the original; however, it fixes a number of
problems with the old implementation.  Currently it does everything the previous
implementation with the exception that it only does the center of charge
calculations because I need to work on how to store the plane information in
this case.  With this new implementation, I've separated items into different
files for better organization with attempted to go through the code.  Everything
for the most part is fully commented.  The command line has changes a slight,
insignificant bit.  Instead of supplying a list of files to examine, the user
just supplies the path to a directory of PDB files or a single PDB file.
This eliminates the need for an extra file containing the filenames of all the
PDB files.  This can be changed back in the future if needed.