Added gz file input capabilities
Included a pretty cool library that can decompress the input PDB files on the
fly. This will eliminate the need to have the user decompress all these files
beforehand thus saving the users hard drive space.  However, this also comes
at the cost of decompressing the files so the runtime could be affected.  The
program can still handle decompressed text files as well so the user has the
option.