Babel support
I have added the ability to add hydrogens using the Babel libraries without
having to write to disk, call Babel, and call another program for further
analysis.  However, there has been talk about not using Babel at all so I
wanted to add this to the repository so that I can have the ability to
rollback if I need to later down the road.