Set up to do angle calcs with hydrogenated residues
Hydrogens are adding correctly and I have function calls set up
to do all of the angle calculations.  I just need the correct equations
to fill in the functions correctly.  Right now, the equations that are
there are simply edited ones from the previous STAAR implementation,
but someone mentioned those we wrong.  Also, added an option to look
for ligands as well, but there is no implementation for doing this
yet.