Center of charge
I know how the center of charge working.  The program also outputs all of the
GAMESS input files into a folder specified by the -g flag.  These files are
named gamessinp-#.inp.  The # is replaced with the number at which it was
found during the STAAR program.  This isn't the most helpful when you are
looking into the directory, but it sure does make it easier when it comes to
submitting jobs as we can just submit an array job (look at the newton
documentation.)  We can look up what these files mean in the main output
table of STAAR.  Now, what's left is to do is new stuff: add ligand support.