Lists and Ligands
Added some lists tho test the program on plus the list that I used
for the Ghost run for the poster.

Added some support for the ligands.  Changed the output file a little
to include titles for the GAMESS results that are added later.  Including
the GAMESS results in the output file will help with creating the database
and web service later