Fix for too many hydrogens
Occassionally, Babel would add too many hydrogens to ASP or GLU residues.
It would add 2 hydrogens to the carbon and 1 to each oxygens. Couldn't
figure out why, but Dr. Hinde and I came up with a fix. What we do is
average the coordinates of the hydrogens that are connected to the carbon.
Then we remove all the hydrogens that Babel added and add this new averaged
coordinates.