Bugs and starts of webservice
There was a bug when looking for interactions between chains. It wouldn't
group the hydrogens correctly before parsing them out into chains if the
residue number was the same.

I also started on a couple of scripts that we will use on the webservice.
These are really basic ones that don't do much on their own.  I have most
of the useful scripts on my own computer since that's where the site is
currently being tested at