Added option to select the PDB resolution
Some of the files had resolutions that were outrageously high and some
that said "NOT APPLICABLE" so I added in the functionality to skip over
PDBs that are above a specified threshold (default: 2 Angstroms) and
ones that say "NOT APPLICABLE".  The command line option to specify
the resolution cut-off is -c or --resolution.