Added support for multiple models
Added support for multiple models, in turn adding support for NMR files.
I also, added a couple of scripts to download PDB files and to make
job submission scripts.  Added a perl script to parse the GAMESS output.
Do not use the C++ version of it because it doesn't work correctly.