Edited the script to download PDB files
It will now download the complete listing of all the new PDB files and
compare them with the listing of the PDB directory.  I will need to change
this later because we won't always have the ability to keep all of the PDB
files.  Instead, we can just keep listings of the PDB id.