Bug fixes
Fixed a bug that was causing STAAR to seg fault if searching inter chain and
the two residues happen to have the same sequence id number on different chains

Fixed a bug that was causing STAAR to seg fault when trying to calculate distances
post hydrogens with residues that had atoms with alt. loc.

Improved the split script that does one mv per split

Changed the script to get new PDB files so that depreciated files are deleted.

Added another filter to dismiss pairs that have distances>threshold after adding
hydrogens.  I was just parsing these out later, but there's no point in
running GAMESS on those if we don't have to.

WIP: Added a new script that runs through the STAAR process (minus downloading PDB)