Fixed altLoc bug
Before, we weren't sure what to do with alternate locations.  Turns out
how we decided to handle them (take the combination of alternat locations
that produced the smallest distance of all combinations) was wrong. Instead,
I changed the code to look at groups of alternate locations (that is, all
atoms with altLoc of 'A' are grouped, 'B' are grouped, etc) and the
one resulting in the smallest distance is used.

Added/edited a bunch of scripts to help with postprocessing and such.